Move Diagnostics carrier to its own module
