+++ src/Control/Effect/Diagnostics.hs
@@ -1,12 +1,6 @@
 module Control.Effect.Diagnostics
   ( Diagnostics (..),
-    DiagnosticsC (..),
     SomeDiagnostic (..),
-    FailureBundle (..),
-    ResultBundle (..),
-    renderFailureBundle,
-    renderWarnings,
-    runDiagnostics,
     fatal,
     context,
     fatalText,
@@ -21,12 +15,7 @@ module Control.Effect.Diagnostics
 where
 
 import Control.Algebra as X
-import Control.Carrier.Error.Either (ErrorC, catchError, runError, throwError)
-import Control.Carrier.Reader (ReaderC, ask, local, runReader)
-import Control.Carrier.Writer.Church (WriterC, runWriter, tell)
 import Control.Exception (SomeException (..))
-import Control.Monad.IO.Class (MonadIO)
-import Data.Monoid (Endo (..))
 import qualified Data.Text as T
 import Data.Text (Text)
 import Data.Text.Prettyprint.Doc
@@ -37,58 +26,6 @@ data Diagnostics m k where
   Fatal :: ToDiagnostic diag => diag -> Diagnostics m a
   Recover :: m a -> Diagnostics m (Maybe a)
   Context :: Text -> m a -> Diagnostics m a
-
-newtype DiagnosticsC m a = DiagnosticsC {runDiagnosticsC :: ReaderC [Text] (ErrorC SomeDiagnostic (WriterC (Endo [SomeDiagnostic]) m)) a}
-  deriving (Functor, Applicative, Monad, MonadIO)
-
-data FailureBundle = FailureBundle
-  { failureWarnings :: [SomeDiagnostic],
-    failureCause :: SomeDiagnostic
-  }
-
-instance Show FailureBundle where
-  show = show . renderFailureBundle
-
-data ResultBundle a = ResultBundle
-  { resultWarnings :: [SomeDiagnostic],
-    resultValue :: a
-  }
-
-renderFailureBundle :: FailureBundle -> Doc AnsiStyle
-renderFailureBundle FailureBundle {..} =
-  vsep
-    [ "----------",
-      "A fatal error occurred:",
-      "",
-      indent 4 (align (renderSomeDiagnostic failureCause)),
-      "",
-      ">>>",
-      "Relevant warnings include:",
-      "",
-      indent 4 (align (renderWarnings failureWarnings))
-    ]
-
-renderSomeDiagnostic :: SomeDiagnostic -> Doc AnsiStyle
-renderSomeDiagnostic (SomeDiagnostic stack cause) = renderDiagnostic cause <> line <> align (indent 2 (vsep (map (pretty . ("when " <>)) stack)))
-
-renderWarnings :: [SomeDiagnostic] -> Doc AnsiStyle
-renderWarnings = align . vsep . map renderSomeDiagnostic
-
-runDiagnostics :: Applicative m => DiagnosticsC m a -> m (Either FailureBundle (ResultBundle a))
-runDiagnostics = fmap bundle . runWriter (\w a -> pure (appEndo w [], a)) . runError @SomeDiagnostic . runReader [] . runDiagnosticsC
-  where
-    bundle (warnings, res) =
-      case res of
-        Left err -> Left (FailureBundle warnings err)
-        Right a -> Right (ResultBundle warnings a)
-
-instance Algebra sig m => Algebra (Diagnostics :+: sig) (DiagnosticsC m) where
-  alg hdl sig ctx = DiagnosticsC $ case sig of
-    L (Fatal diag) -> ask >>= \path -> throwError (SomeDiagnostic path diag)
-    L (Context path go) -> local (path :) $ runDiagnosticsC $ hdl (go <$ ctx)
-    L (Recover act) -> do
-      (fmap (fmap Just)) (runDiagnosticsC $ hdl (act <$ ctx)) `catchError` (\(diag :: SomeDiagnostic) -> tell (Endo (diag :)) *> pure (Nothing <$ ctx))
-    R other -> alg (runDiagnosticsC . hdl) (R (R (R other))) ctx
 
 fatal :: (Has Diagnostics sig m, ToDiagnostic diag) => diag -> m a
 fatal = send . Fatal
